test problem
Tested this testcall.dump, all lines except for the last two behave as
expected (I haven't checked the formatting yet, though). lui gives a
zero answer, and slt is not executed (I assume this is because the
simulator only runs a set number of commands, but haven't confirmed that
yet).